+++ tket2-py/tket2/circuit/__init__.py
@@ -1,0 +1,6 @@
+# Re-export native bindings
+
+from ..tket2 import _circuit
+from ..tket2._circuit import *  # noqa: F403
+
+__all__ = _circuit.__all__
